Fix module imports when building with Qt Version 6.0.0
For modules such as QtQuick we register 2.$QT_VERSION_MINOR to allow our
users to simply use the latest "Qt version" with their imports. When
switching to version 6.0.0, code that used import QtQuick 5.13 would
stop working because we had some types registered to say minor version 10
but nothing after that, and the qmlRegisterModule() call would use
QT_VERSION_MINOR, which is now zero. Therefore in this Qt 6 branch,
let's stick to the latest Qt 5 release planned, to maintain
compatibility.

Change-Id: I174be32cef8de152cd32010156ae716bd474397c
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>