Fix import of ListModel/ListElement as part of QtQml
ListElement and ListModel were available in Qt 5 since minor version 0
in QtQml and since minor version 1 in QtQml.Models. So one had to either
write "import QtQml 2.0" or "import QtQml.Models 2.1".

Since the explicit registration in QtQml is gone and to be replaced with
an implicit dependency that loads QtQml.Models when loading QtQml, with
the same minor version, we need to lower the minor version to 0 for both.

This is unlikley to cause any harm as previously import QtQml.Models 2.0
would not produce any types.

Change-Id: Iff7c4a1c1b35e4c9d930aa45dd678d8e2d5c823c
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>