Fix deployment of QtQuick for Android with dummy_imports.qml
We rely on the qmlimportscanner to find out what to deploy, but it is
limited to scanning .qml files. These tests creates the QML code from C++,
so qmlimportscanner fails to detect those dependencies.

Therefore, we add a dummy_imports.qml file that has the sole purpose of
giving qmlimportscanner which additional dependencies the test has.

Change-Id: I4237b738e408c309b9b21de7e53d1a4e4acb7e2e
Task-number: QTBUG-73512
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>