Remove module registration related Qt version #ifdefs
Ahead of compiling declarative with a newer qtbase that bumps the Qt
version to 6, we can already now remove the Qt 5 code paths related to
type registrations as commit 8534634bd204ccdca3edcc244a35be6e81739fce
implements those dependencies differently.

This also required lowering the minimum minor version for the
Instantiator to 1, as it was in the Qt 5 branch.

Change-Id: Idd38c967eeb8423509237116803d292bba422f8b
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>