Fix bug in QQmlEngine::setImportPathList() when it had a resource path
The bug is really in QQmlImportDatabase::setImportPathList(). It was
missing the same conversions that was done in
QQmlImportDatabase::addImportPath(), so it failed to use a resource path
as a import path because it did not convert ":/foo" to "qrc:/foo".

We therefore just use addImportPath() to ensure the paths are converted
properly.

Before this, several autotests in tests/auto/qml/qqmllanguage failed on
Android, since they were calling QQmlEngine::setImportPathList() where the
list had resource paths.

Task-number: QTBUG-73512
Change-Id: Idc64f5ad20ec665df7cb57ea1c346bc0975c3b0d
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>