Implement QSGRenderNode for the rhi path
The documentation is also refined and extended.

Revise the rendernode example as well: make it possible to
test and demonstrate both scissor and stencil based clipping.

An implementation of the triangle for another graphics API
should follow in a separate patch at a later point. For now
only OpenGL is supported in combination with the RHI.

Change-Id: I7ef9f6c6e0b44f1bdf44c9266ea3fa4736367a5d
Reviewed-by: Christian Strømme <christian.stromme@qt.io>