Allow Connections to handle signals using JavaScript functions
Requiring full function definitions as signal handlers has two
advantages:

1, We don't need a custom parser that magically recognizes properties
   which would otherwise be an error in other components.
2, The user is forced to specify the full signature of the handler,
   including any parameters. This helps when the functions will
   eventually be compiled to C++

The old behavior is retained, generating a warning if any of the magic
bindings are still set in a Connections element. Only if no magic
bindings are found, the functions are connected. This is because there
might be functions named onFoo in old-style Connections elements and
silently connecting those to any matching signals would be a change in
behavior.

Change-Id: I8c78d8994fdcddd355fe822cde9a0702dc8c75de
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>