Generate registrations for all imports
Now that we can generate all QML type information at build time, we
should also use it.

Move the types for QtTest into the testlib import plugin. They don't
need to be exported from QtQuickTest.

Trigger the resource initialization of the shapes library from the
plugin so that we retain a dependency and the linker doesn't optimize
the module initialization away.

Change-Id: Icc8cb338aa03ef1e3085e29356a5db7b73ba0a01
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>