Fix not printing logs with QSG_INFO when forcing rhi backend from C++
QQuickWindow::setScenegraphBackend() may be called before QQuickWindow
gets a chance to create a render loop. If QSG_INFO (the env.var.) is
used instead of the logging category (qt.scenegraph.general), some logs
are not printed because the code that enables the logging category is not
yet run.

To prevent confusion, make sure the logging category gets enabled before
the first potential qCDebug. In the worst case we check twice but that's
fine.

Change-Id: Ibfc0af05050adc9766c30a2d15c778b2a51823fe
Reviewed-by: Andy Nichols <andy.nichols@qt.io>