qqmljs.g: reduce number of conflicts
Use right associativity to tell the parser to shift directly when
encountering an else statement. This is the wanted behavior, as the
"else" should be associated with the last "if" statement.
This removes one more shift/reduce conflict in the grammar.

See also https://stackoverflow.com/a/12734499 (qlalr behaves like
Bison/Yacc here).

Change-Id: Id29f00b5a2862fdb3399e4c8a183eda2ed328ba4
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>