Avoid introducing rhiTexture() in the public API of QSGTexture
There is no need for this. The few internal users can get it via
QSGTexturePrivate.

The original thinking was based on QRhi* being a public API, but that
is not the case in the near future. So avoid introducing a public API
relying on QRhiTexture.

This of course makes it impossible to retrieve the native object under
a QSGTexture (as textureId() is not used anymore when rendering with the
RHI). For that, an alternative approach will be introduced later on.

Change-Id: I0099b23424cafa4958f78c03300b0c934b60d92c
Reviewed-by: Christian Strømme <christian.stromme@qt.io>