Rename textureinsgnode example to a more fitting name
It shows QQuickFramebufferObject. "Texture in QSGNode" is ambiguous
as it suggests another type of use case (where an existing texture is
imported and used with the QQ scene - without rendering to it). This
would conflict with upcoming future examples, so rename to fboitem, which
is what this example is about.

The link from examples.qdoc is already using the fitting title so that will
just work.

Change-Id: Ide7a563398a4b464c55f14e10fa9386432b3d06d
Reviewed-by: Andy Nichols <andy.nichols@qt.io>