Allow semicolon after property declaration
Most of the rules already had Semicolon at the end, however it was
missing for UiScriptStatement, list properties and UiObjectInitializer.

This change fixes the regression from 5.11.3 to 5.12.0, and keeps the
behavior consistent.

(Semicolon was only introduced in 5.14, that why we need to introduce
the rule in 5.12 first)

Fixes: QTBUG-77954
Change-Id: I45ef35fab399e3f971444b96d4a9ec6a99e29e09
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit 45b1a3f97953fac65c6aef8e46abad865a0d0bc3)