Add qt6_add_qml_module() public API
Add qt6_add_qml_module() as a public API for building QML modules.

Since the shared implementation details are small, it was easier
to just reimplement the shared code than to unmangle the more
complicated version we use to build Qt from QtBuild.

This patch includes an example conversion. Changes for pro2cmake
will follow in a separate patch in qtbase.

Change-Id: I942526cc7d978e2d8309b506e785f9c1509d0bbc
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>