Prospective fix for the Windows build
In qmltest.pro the MODULE variable is not set, so the module name is
derived from the name of the .pro file. The same discrepancy exists in
qtbase's QtTest library where the module name is QtTestLib. There we've
special cased it in the CMakeLists.txt and so this patch apples the same
logic here.

Change-Id: I9678cfb8fe407a72be927183a1eca47d0f68ffc3
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>