QQmlTypeLoader: Prevent trivial cycles
The type loader already contained a cycle check, however it did not work
for the attached example. We now do an additional test for A->B->A
cycles, which fixes the bug. This is a valid strategy, as in that case
the currently resolved type would immediately fullfill the depentency of
the type waiting on it, which in turn resolves the only dependency of
the current type.

Fixes: QTBUG-78098
Change-Id: I8b550a1c240d3d0fbf158beb99a5daf992904cb0
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>