Convert QTime to V4 Date on a date with no DST in force
Sadly the third centennary of Cassini's birth isn't such a good date
after all, since round-tripping will ignore DST on the QDateTime leg
but include it in V4's implementation of ES's ill-conceived spec for
handling DST and time-zone transitions; and June 8th is in summer for
much of the world. So use a date which (as far as I can tell) appears
to have been DST-free everywhere; and after the start of 1970 so that,
even if we find ourselves working with a zone in which it did have
DST, we'll apply DST consistently during round-tripping.

Fixes: QTBUG-78706
Change-Id: I25be2548e2c97eae2c4f6c5bbe661dea2f35aaba
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>