WorkerScript: Directly delete QQuickWorkerScriptEnginePrivate
QQuickWorkerScriptEnginePrivate lives in the worker thread. Therefore,
once the thread has finished there is no way to send it a deferred
delete event. The object and all its children would always leak. As
there is no event loop running in the worker thread anymore and this is
the dtor of QQuickWorkerScript, it's safe to assume that no one can
access the private object anymore afterwards.

Change-Id: I51f583ea47060d967403639196247882ff7d2905
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>