QQmlObjectModel: Return object property from variantValue()
Returning a context property was a strange choice. We want to get rid of
context properties.

[ChangeLog][Important Behavior Changes] The QML ObjectModel type had an
undocumented "feature" which made it sort items in list views into
sections according to context properties of the QML context the created
objects belong to. Instead of context properties, object properties are
used now.

Change-Id: Ibd94c903910f0cdb76ead50db92a24eefebfa3fc
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>