VDMAbstractItemModelDataType: ensure meta-type is initialized
Since the delegate chooser can call value() on a delegate model
before the first call to createItem(), we need to ensure
that all structures (and in particular, m_rolenames) are
populated also in this case.

This is especially important for cases where value() is called
for every item in the model, as is the case for e.g. auto-completion
for combo boxes, as delegate creation is particularly expensive
there.

Task-number: QTBUG-78858
Change-Id: I66138d74839b9645d11c69c29f2e70264c65a68d
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>